refactor(CodeBlocks): split `Code` into 3 separate components
issue #776
