+++ components/CodeBlocks/InlineCode.module.css
@@ -1,4 +1,4 @@
-.inline {
+.code {
   /* Add `!important` to override `antd` styles */
   @apply px-1.5 py-px !important;
   @apply border border-solid border-secondary !important;
@@ -9,16 +9,4 @@
   @apply shadow-xl !important;
 
   font-size: var(--font-size) !important;
-}
-
-.block {
-  @apply static !important;
-  @apply mb-0 p-5 !important;
-}
-
-.line-number {
-  @apply inline-block !important;
-  @apply w-6 !important;
-  @apply mr-3 !important;
-  @apply text-secondary select-none !important;
 }
