+++ components/CodeBlocks/LiveCode.module.css
@@ -1,0 +1,4 @@
+.code {
+  @apply relative;
+  @apply m-0 p-5;
+}
