+++ components/CodeBlocks/Pre.tsx
@@ -1,53 +1,24 @@
 import { classNames } from '@components/utils';
 import type { HTMLProps, ReactElement } from 'react';
-import Code from './Code';
+import BlockCode from './BlockCode';
+import LiveCode from './LiveCode';
 import styles from './Pre.module.css';
+import { normalizeLanguage } from './utils';
 
 interface Props extends HTMLProps<HTMLPreElement> {
+  live?: boolean;
+  noline?: boolean;
+  nocopy?: boolean;
   title?: string;
 }
 
-const normalizeLanguage = (language: string) => {
-  switch (language) {
-    case 'html':
-      return 'HTML';
-    case 'xml':
-      return 'XML';
-    case 'yml':
-    case 'yaml':
-      return 'YAML';
-    case 'css':
-      return 'CSS';
-    case 'json':
-      return 'JSON';
-    case 'md':
-    case 'markdown':
-      return 'Markdown';
-    case 'js':
-    case 'javascript':
-      return 'JavaScript';
-    case 'ts':
-    case 'typescript':
-      return 'TypeScript';
-    case 'coffee':
-    case 'coffeescript':
-      return 'CoffeeScript';
-    case 'jsx':
-    case 'tsx':
-      return 'React';
-    case 'objc':
-    case 'objectivec':
-      return 'Objective-C';
-    default:
-      return language?.charAt(0)?.toUpperCase() + language?.slice(1) || 'Code';
-  }
-};
-
 const Pre = ({
+  live = false,
+  noline = false,
+  nocopy = false,
   title,
   children,
   className,
-  ...props
 }: Props): JSX.Element => {
   const codeElement = children as ReactElement;
   const code: string = codeElement?.props?.children;
@@ -61,9 +32,17 @@ const Pre = ({
       className={classNames(className, languageClass, styles.pre)}
       data-language={title || languageName}
     >
-      <Code {...props} className={languageClass}>
-        {code}
-      </Code>
+      {live ? (
+        <LiveCode className={languageClass}>{code}</LiveCode>
+      ) : (
+        <BlockCode
+          enableLine={!noline}
+          enableCopy={!nocopy}
+          className={languageClass}
+        >
+          {code}
+        </BlockCode>
+      )}
     </pre>
   );
 };
