+++ components/CodeBlocks/index.ts
@@ -1,2 +1,2 @@
-export { default as Code } from './Code';
+export { default as InlineCode } from './InlineCode';
 export { default as Pre } from './Pre';
