+++ components/MDX/MDXCode.tsx
@@ -1,7 +1,7 @@
-import { Code } from '@components/CodeBlocks';
+import { InlineCode } from '@components/CodeBlocks';
 
 interface Props {}
 
-const MDXCode = (props: Props) => <Code {...props} />;
+const MDXCode = (props: Props) => <InlineCode {...props} />;
 
 export default MDXCode;
